timezones-alternative: fix do_install

+++ packages/timezones/timezones-alternative.bb
@@ -7,13 +7,14 @@ LICENSE = "GPL"
 SRC_URI = "http://sources.dreamboxupdate.com/zoneinfo.tar.bz2"
 S = "${WORKDIR}/zoneinfo"
 
+FILES_${PN} = "usr/share/zoneinfo/[A-Z]*"
+
 do_install() {
 	install -d ${D}/usr/share/zoneinfo/
 	
 	for file in ${S}/*
 	do
-		install -m 644 "$file" ${D}/usr/share/zoneinfo/
-	done
+		[ -f $file ] && install -m 644 "$file" ${D}/usr/share/zoneinfo/
+	done;
+	true;
 }
-
-FILES_${PN} = "usr/share/zoneinfo/[A-Z]*"
